+++ documentation/docs/user/releasenotes.md
@@ -3,7 +3,8 @@
 ## Version 0.2.0
 - Added Skip action to skip the first 'n' data items of a stream.
 - Added Limit action to filter data items from the stream after the first 'n' data items have been processed.
-- 'Fixed input' provider now supports JSON as input.
+- 'Fixed Input' provider now supports JSON as input.
+- Added 'Split Array' action to split an input array into multiple data items.
 
 ## Version 0.1.0
 - First stable release with core features.